backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
bcrypt
PyJWT
pydantic
pydantic-settings
email-validator
python-multipart
google-genai
pydantic>=2.0.0
